% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{HeckmanEM.criteria}
\alias{HeckmanEM.criteria}
\title{Model selection criteria for the Heckman Selectio model}
\usage{
HeckmanEM.criteria(obj)
}
\arguments{
\item{obj}{An object of the class HeckmanEM.}
}
\value{
The calculated AIC, AICc, and BIC for the parameters of the fitted model.
}
\description{
`HeckmanEM.criteria()` calculate the AIC, AICc, BIC selection criteria for the fitted Heckman selection model.
}
\examples{
\dontrun{
res <- HeckmanEM(y, x, w, cc, nu = 4, family = "Normal", error = 1e-05, iter.max = 500, 
                 im = TRUE, criteria = FALSE)
cr <- HeckmanEM.criteria(res)
}
}
