% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.R
\name{HeckmanEM.infomat}
\alias{HeckmanEM.infomat}
\title{Standard error estimation for the Heckman Selectio model by the Information Matrix.}
\usage{
HeckmanEM.infomat(obj)
}
\arguments{
\item{obj}{An object of the class HeckmanEM.}
}
\value{
The estimated standard errors for the parameters of the fitted model.
}
\description{
`HeckmanEM.infomat()` estimate the standard errors for the parameters for the fitted Heckman selection model.
}
\examples{
\dontrun{
res <- HeckmanEM(y, x, w, cc, nu = 4, family = "Normal", error = 1e-05, iter.max = 500, 
                 im = FALSE, criteria = TRUE)
im <- HeckmanEM.infomat(res)
}
}
