% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diag_Heckman-t.R
\name{Influence}
\alias{Influence}
\title{Influence Analysis for the Heckman Selection model}
\usage{
Influence(object, type, colx = NULL, k = 3.5)
}
\arguments{
\item{object}{A `HeckmanEM` object to perform the analysis on.}

\item{type}{A character string indicating the type of perturbation to perform.
The types can be one of "case-weight","scale","response" and"exploratory".}

\item{colx}{Optional integer specifying the position of the column in the
object's matrix \code{x} that will undergo perturbation. Only required when type is
"exploratory".}

\item{k}{A positive real constant to be used in the benchmark calculation: \eqn{M_0 + k\times \mathrm{sd}(M_0)}. Default is 3.5.}
}
\value{
Returns a list with the following elements:
\item{M0}{A vector of length \eqn{n} with the aggregated contribution of all eigenvectors of the matrix associated with the normal curvature.}
\item{benchmark}{\eqn{M_0 + k\times \mathrm{sd}(M_0)}}
\item{influent}{A vector with the influential observations' positions.}
}
\description{
This function conducts influence analysis for a given `HeckmanEM` object.
The influence analysis can be conducted using several types of perturbations.
}
\examples{
\dontrun{
# Assume `heckman_model` is a HeckmanEM object
Influence(heckman_model, type = "case-weight")
}

}
\references{
Insert any relevant references here.
}
\seealso{
\code{\link{HeckmanEM}}
}
\author{
Marcos Oliveira
}
